+++ src/mobgap/laterality/_lrc_benmansour.py
@@ -1,9 +1,8 @@
-from typing import Any
+from typing import Any, Self, Unpack
 
 import numpy as np
 import pandas as pd
 from tpcp import cf
-from typing_extensions import Self, Unpack
 
 from mobgap.data_transform import ButterworthFilter
 from mobgap.data_transform.base import BaseFilter